+++ src/main/java/roomescape/exception/customexception/ReservationNotFoundException.java
@@ -1,6 +1,6 @@
-package roomescape.exceptions;
+package roomescape.exception.customexception;
 
-import roomescape.ErrorCode;
+import roomescape.exception.model.ErrorCode;
 
 public class ReservationNotFoundException extends CustomException {
     public ReservationNotFoundException() {
